libs/lib_agent/pyproject.toml
[project]
name = "lib_agent"
version = "0.0.0"
description = ""
requires-python = ">=3.13,<3.14"
authors = []
dependencies = [
    "discrete-dists~=1.0",
    "chex>=0.1.89",
    "distrax>=0.1.5",
    "dm-haiku>=0.0.14",
    "flax>=0.10.6",
    "jax>=0.6.1",
    "lib_utils[jax]",
    "lib_config",
    "optax>=0.2.4",
]
readme = "README.md"

[dependency-groups]
dev = [
    "pylint>=3.3.7",
    "pyright>=1.1.392.post0",
    "pytest>=8.4.0",
    "pytest-benchmark>=5.1.0",
    "pytest-xdist>=3.7.0",
    "ruff>=0.11.12",
    "uv>=0.5.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool]
[tool.hatch.build.targets.wheel]
packages = ["lib_agent"]

[tool.ruff]
extend = "../../standards/python/ruff.toml"

[tool.pyright]
extends='../../standards/python/pyright.toml'
include = ['.']
exclude = [
    '.venv/',
]
venvPath = '.'


[tool.commitizen]
name = "cz_conventional_commits"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true

[tool.uv.sources]
lib_utils = { path = "../lib_utils", editable = true }
lib_config = { path = "../lib_config", editable = true }
